projects/api/scripts/requirements.txt
selenium==4.35.0
PyYAML==6.0.1
requests==2.32.4
mutagen==1.47.0
cairosvg>=2.7.0
black>=24.0.0
isort>=5.13.0
pylint>=3.0.0
flake8>=7.0.0
pip-audit>=2.10.0
types-PySocks>=1.7.1.20251001
types-PyYAML>=6.0.12.20250915
types-requests>=2.32.4.20250913
mypy>=1.19.1